outsource RelationStatisticsMixin (from RETextClassificationWithIndicesTaskModule)
