+++ src/pie_modules/taskmodules/common/mixins.py
@@ -1,9 +1,15 @@
 import dataclasses
-from typing import Any, Dict, Iterable, List, Optional
+import logging
+from collections import defaultdict
+from typing import Any, Dict, Iterable, List, Optional, Tuple
 
+import pandas as pd
 import torch
 import torch.nn.functional as F
+from pytorch_ie import Annotation
 from torch import Tensor
+
+logger = logging.getLogger(__name__)
 
 
 def _pad_tensor(tensor: Tensor, target_shape: List[int], pad_value: float) -> Tensor:
@@ -143,3 +149,65 @@ class BatchableMixin:
             # Only batch attributes that are not None for any of the values.
             if not all(getattr(x, k) is None for x in values)
         }
+
+
+class RelationStatisticsMixin:
+    def __init__(self, collect_statistics: bool = False, **kwargs):
+        super().__init__(**kwargs)
+        self.collect_statistics = collect_statistics
+        self.reset_statistics()
+
+    def reset_statistics(self):
+        self._statistics = defaultdict(int)
+        self._collected_relations: Dict[str, List[Annotation]] = defaultdict(list)
+
+    def collect_relation(self, kind: str, relation: Annotation):
+        if self.collect_statistics:
+            self._collected_relations[kind].append(relation)
+
+    def collect_all_relations(self, kind: str, relations: Iterable[Annotation]):
+        if self.collect_statistics:
+            self._collected_relations[kind].extend(relations)
+
+    def finalize_statistics(self):
+        if self.collect_statistics:
+            all_relations = set(self._collected_relations["available"])
+            used_relations = set(self._collected_relations["used"])
+            skipped_other = all_relations - used_relations
+            for key, rels in self._collected_relations.items():
+                rels_set = set(rels)
+                if key.startswith("skipped_"):
+                    skipped_other -= rels_set
+                elif key.startswith("used_"):
+                    pass
+                elif key in ["available", "used"]:
+                    pass
+                else:
+                    raise ValueError(f"unknown key: {key}")
+                for rel in rels_set:
+                    self.increase_counter(key=(key, rel.label))
+            for rel in skipped_other:
+                self.increase_counter(key=("skipped_other", rel.label))
+
+    def show_statistics(self):
+        if self.collect_statistics:
+            self.finalize_statistics()
+
+            to_show = pd.Series(self._statistics)
+            if len(to_show.index.names) > 1:
+                to_show = to_show.unstack()
+            to_show = to_show.fillna(0)
+            if to_show.columns.size > 1:
+                to_show["all_relations"] = to_show.loc[:, to_show.columns != "no_relation"].sum(
+                    axis=1
+                )
+            if "used" in to_show.index and "available" in to_show.index:
+                to_show.loc["used %"] = (
+                    100 * to_show.loc["used"] / to_show.loc["available"]
+                ).round()
+            logger.info(f"statistics:\n{to_show.to_markdown()}")
+
+    def increase_counter(self, key: Tuple[Any, ...], value: Optional[int] = 1):
+        if self.collect_statistics:
+            key_str = tuple(str(k) for k in key)
+            self._statistics[key_str] += value
